Conflicts:
	Library/PureLib/Utility/WindowConsole.cs
	Projects/FishHunter/Formula/Formula.cs
	Projects/FishHunter/Game/Formula/ZsFormula/ZsFishStage.cs
	Projects/FishHunter/StorageServer/Server.cs

+++ Library/PureLib/Utility/PowerRegulator.cs
@@ -41,7 +41,7 @@
 			_FPS = new FPSCounter();
 		}
 
-		public void Operate(int busy)
+		public void Operate(long busy)
 		{
 			_FPS.Update();
